# openapi.yaml
openapi: 3.0.3
info:
  title: Smart Attendance & Classroom Verification API (ATT)
  version: 1.1.0
  description: |
    Backend API for the Smart Attendance project (requirement prefix ATT).

    **Authentication:** call `POST /auth/login`, then send the returned token in every request as
    `Authorization: Bearer <token>`. Tokens are valid for 8 hours.

    **Roles:** `student`, `lecturer`, `ta`, `admin`, `auditor`. Authorization is enforced on the backend.

    **QR tokens:** the lecturer/TA app calls `GET /sessions/{id}/qr` repeatedly (about every 15 seconds).
    Each token expires after 20 seconds. The student app sends the scanned token to `POST /attendance/scan`.

    **Dates and times:** dates are `YYYY-MM-DD`, timestamps are ISO 8601 (UTC), times are `HH:MM:SS`.

    **Rate limiting:** `POST /attendance/scan` allows 10 attempts per minute per student account
    (HTTP 429 with `reason: too_many_attempts` afterwards).

    **Flags:** `GET /flags` returns rule-based, advisory-only attendance flags. They never change
    attendance and can be unavailable (HTTP 503) without affecting attendance capture.

    All data used in this project is synthetic.
servers:
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /sessions/my-slots:
    get:
      tags: [Sessions]
      summary: List my scheduled slots (lecturer/TA)
      description: |
        Returns the timetable slots of every section the caller is assigned to.
        `is_today` is true when the slot's weekday is today (Cairo time).
        `session_id` and `session_status` are filled only if a session was already opened today.
      responses:
        '200':
          description: List of slots
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    slot_id:
                      type: integer
                    course_code:
                      type: string
                    course_name:
                      type: string
                    section_id:
                      type: integer
                    section_name:
                      type: string
                    room_name:
                      type: string
                    building:
                      type: string
                      nullable: true
                    day_of_week:
                      $ref: '#/components/schemas/DayOfWeek'
                    start_time:
                      type: string
                      example: '10:00'
                    end_time:
                      type: string
                      example: '12:00'
                    session_id:
                      type: integer
                      nullable: true
                    session_status:
                      type: string
                      nullable: true
                      enum: [open, closed]
                    is_today:
                      type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /sessions/open:
    post:
      tags: [Sessions]
      summary: Open an attendance session (lecturer/TA)
      description: |
        Only staff assigned to the section can open its scheduled slot, only on the slot's weekday,
        and only inside the slot's time range. Opening freezes the roster: every active enrolled
        student gets an `absent` attendance record.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slot_id]
              properties:
                slot_id:
                  type: integer
      responses:
        '201':
          description: Session opened
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenedSession'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /sessions/{id}/close:
    post:
      tags: [Sessions]
      summary: Close an open session (lecturer/TA who opened it)
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session closed
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: integer
                  status:
                    type: string
                    example: closed
                  closed_at:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /sessions/{id}/qr:
    get:
      tags: [Sessions]
      summary: Get the current rotating QR token (lecturer/TA who opened the session)
      description: Every call returns a new signed token that expires after 20 seconds. Render it as a QR code.
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: QR token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  expires_in_seconds:
                    type: integer
                    example: 20
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SessionId:
      name: id
      in: path
      required: true
      description: Attendance session id
      schema:
        type: integer
...
  responses:
    BadRequest:
      description: Missing or invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing, invalid or expired token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The role is not allowed, or the record belongs to someone else
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Record not found (or not yours)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Duplicate record, or the action conflicts with the current state (e.g. room already booked, open session)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    DayOfWeek:
      type: string
      enum: [Saturday, Sunday, Monday, Tuesday, Wednesday, Thursday, Friday]
...
    OpenedSession:
      type: object
      properties:
        session_id:
          type: integer
        slot_id:
          type: integer
        session_date:
          type: string
          format: date
        status:
          type: string
          enum: [open, closed]
        opened_at:
          type: string
          format: date-time
        roster_count:
          type: integer
          description: Number of enrolled students added to the roster
...